Removed Transition data structure, embed the transition information into Node instead. This helps significantly reduce memory usage.
Also to properly implement the longest common substring method, a unique terminator character is inserted to the end of every string. This now puts the limitation that we can only support at most of a fixed number of strings in the tree.

The longest common substring algorithm is reimplemented properly using least common ancestor (Tarjan's algorithm).

Moved tests out to a separate folder.